#13 cambio variable poblacion_adulta a uint16

+++ pkg/segurasenior/poblacion.go
@@ -6,14 +6,14 @@ type Datos_poblacion struct {
 	Nombre_pueblo       string
 	Poblacion_total     uint16
 	Poblacion_joven     uint16
-	Poblacion_adulta    int
+	Poblacion_adulta    uint16
 	Poblacion_anciana   int
 	Tasa_natalidad      float64
 	Tasa_envejecimiento float64
 	Tasa_mortalidad     float64
 }
 
-func NewPoblacion(nom string, poblacion uint16, poblacionjoven int, poblacionmedia int, poblacionmayor int, tasan float64, tasae float64, tasam float64) (*Datos_poblacion, error) {
+func NewPoblacion(nom string, poblacion uint16, poblacionjoven uint16, poblacionmedia uint16, poblacionmayor int, tasan float64, tasae float64, tasam float64) (*Datos_poblacion, error) {
 	if nom == "" {
 		return nil, errors.New("nombre no puede estar vacío")
 	}
